Spotify: correct mistakes of last task
