+++ Spotify/Utils/Extensions/UIView+Extensions.swift
@@ -40,26 +40,3 @@ extension UIView {
         }
     }
 }
-
-extension DateFormatter {
-    static let dateFormatter: DateFormatter = {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "YYYY-MM-dd"
-        return dateFormatter
-    }()
-
-    static let displayDateFormatter: DateFormatter = {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateStyle = .medium
-        return dateFormatter
-    }()
-    
-        static func formattedDate(string: String) -> String {
-            guard let date = DateFormatter.dateFormatter.date(from: string) else {
-                return string
-            }
-            return DateFormatter.displayDateFormatter.string(from: date)
-        }
-    }
-
-
